category test

+++ app/Http/Controllers/CategoryController.php
@@ -14,21 +14,30 @@ use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 class CategoryController extends Controller
 {
     use AuthorizesRequests;
+    private $imageService;
+
+    public function __construct(ImageService $imageService)
+    {
+        $this->imageService = $imageService;
+    }
 
     public function index()
     {
+        $this->authorize('viewAny', Category::class);
         $categories = Category::all();
         return view('category.index', compact('categories'));
     }
 
     public function create()
     {
+        $this->authorize('create', Category::class);
         $categories = Category::all();
         return view('category.create', compact('categories'));
     }
 
     public function store(StoreRequest $request, ImageService $imageService, CkeditorService $editorService)
     {
+        $this->authorize('viewAny', Category::class);
         $data = $request->validated();
 
         if ($request->hasFile('image')) {
@@ -46,6 +55,7 @@ class CategoryController extends Controller
     public function edit($slug)
     {
         $category = Category::where('slug', $slug)->first();
+        $this->authorize('update', $category);
 
         if (!isset($category)) {
             return redirect()->route('category.index')->with('error', 'دسته بندی وجود ندارد');
@@ -57,6 +67,7 @@ class CategoryController extends Controller
     public function update(UpdateRequest $request, ImageService $imageService, CkeditorService $editorService, $slug)
     {
         $category = Category::where('slug', $slug)->first();
+        $this->authorize('update', $category);
         if (!isset($category)) {
             return redirect()->route('category.index', 'دسته بندی یافت نشد');
         }
@@ -78,9 +89,21 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         $category = Category::find($id);
+        $this->authorize('delete', $category);
 
         if (!isset($category)) {
             return back()->with('error', 'دسته بندی وجود ندارد');
+        }
+
+        //         // حذف عکس اصلی
+        if ($category->image) {
+            $this->imageService->delete($category->image);
+        }
+
+        // حذف تصاویر CKEditor
+        foreach ($category->editorImages as $editorImage) {
+            $this->imageService->delete($editorImage->image);
+            $editorImage->delete();
         }
 
         $category->delete();
